Improve content touch interception (Fixes GH-9)
Instead of recursively enabling/disabling the content ViewGroup
when the drawer closes/opens, just do a simple check in
onInterceptTouchEvent to see if the touch needs to be
intercepted on an ACTION_UP event.

Also intercept ACTION_DOWN events when the drawer is in motion
and the event originated within the content ViewGroup.

Signed-off-by: Eddie Ringle <eddie@eringle.net>